requirements.txt
openai==0.28
matplotlib
python-dotenv
requests==2.31.0